Converted the habits to a two-column layout.
